project/ocr-service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
paddlepaddle==2.6.0
paddleocr==2.7.3
Pillow==10.1.0
numpy==1.26.2
opencv-python-headless==4.8.1.78
pdf2image==1.16.3
pypdf==3.17.1
